Add files via upload
This is a .zip file that contains our final data mining project. We used five different classifiers for our porject to find which would be the best one. The file contains the markdown and .ipynb files for the five classifiers as well as the data preprocessing files. The folder labeled "saved_models" is a directory that holds .pkl files that contain trained models that took a long time to compute. The "Images" folder contains all the visuals from all the .ipynb files. The "Data" folder contains the raw and preprocessed data. This project was done using python 3.6.5.